[Dev-264] MQTT: Add an ability to connect to a broker without user credentials [Dev-263] MQTT: Add Redis store as a persistence for a broker Integration test improvements Implementation of Redis persistence storage
