+++ lib/SubscriptionManager.js
@@ -3,6 +3,7 @@ const SubscriptionEntry = require(`./SubscriptionEntry.js`);
 
 /**
  * Subscription Manager class. Used to handle multiple subscription
+ * @event message
  */
 class SubscriptionManager extends EventEmitter {
