+++ lib/WebSocket.js
@@ -1,17 +1,19 @@
 const EventEmitter = require(`events`);
 const WS = require(`ws`);
-const randomstring = require(`randomstring`);
 
 const MAX_LISTENERS = 20;
 
 
 /**
- *
+ * WebSocket class wrapper
+ * @event open
+ * @event close
+ * @event message
  */
 class WebSocket extends EventEmitter {
 
     /**
-     *
+     * Create WebSocket
      * @param serverURL
      */
     constructor(serverURL) {
@@ -36,7 +38,7 @@ class WebSocket extends EventEmitter {
     }
 
     /**
-     *
+     * Close WebSocket session
      */
     close() {
         const me = this;
@@ -45,7 +47,7 @@ class WebSocket extends EventEmitter {
     }
 
     /**
-     *
+     * Add event listener for WebSocket
      * @param eventName
      * @param listener
      */
@@ -56,7 +58,7 @@ class WebSocket extends EventEmitter {
     }
 
     /**
-     *
+     * Remove event listener for WebSocket
      * @param eventName
      * @param listener
      */
@@ -67,48 +69,7 @@ class WebSocket extends EventEmitter {
     }
 
     /**
-     *
-     * @param messageObject
-     * @return {Promise}
-     */
-    send(messageObject) {
-        const me = this;
-
-        return new Promise((resolve, reject) => {
-            const hasRequestId = messageObject.requestId;
-
-            if (!hasRequestId) {
-                messageObject.requestId = randomstring.generate();
-            }
-
-            me.ws.send(JSON.stringify(messageObject));
-
-            const listener = (event) => {
-                const messageData = JSON.parse(event.data);
-
-                if (messageData.action === messageObject.action) {
-                    if (messageData.requestId === messageObject.requestId) {
-                        if (!hasRequestId) {
-                            delete messageData.requestId;
-                        }
-
-                        me.ws.removeListener(`message`, listener);
-
-                        if (messageData.status === `success`) {
-                            resolve(messageData);
-                        } else {
-                            reject(messageData);
-                        }
-                    }
-                }
-            };
-
-            me.ws.addEventListener(`message`, listener);
-        });
-    }
-
-    /**
-     *
+     * Send data as string
      * @param str
      * @param callback
      */
