+++ lib/WebSocketFactory.js
@@ -2,12 +2,13 @@ const EventEmitter = require('events');
 const WebSocket = require('./WebSocket.js');
 
 /**
- *
+ * WebSocket Factory. For supporting multiple users
+ * @event message
  */
 class WebSocketFactory extends EventEmitter {
 
     /**
-     *
+     * Create WebSocketFactory
      * @param serverUrl
      */
     constructor(serverUrl) {
@@ -20,9 +21,9 @@ class WebSocketFactory extends EventEmitter {
     }
 
     /**
-     *
+     * Get WebSocket by key or create new WebSocket
      * @param key
-     * @return {*}
+     * @return {Promise}
      */
     getSocket(key) {
         const me = this;
@@ -43,7 +44,7 @@ class WebSocketFactory extends EventEmitter {
     }
 
     /**
-     *
+     * Remove WebSocket by key
      * @param key
      */
     removeSocket(key) {
@@ -56,7 +57,7 @@ class WebSocketFactory extends EventEmitter {
     }
 
     /**
-     *
+     * Check for socket with such key
      * @param key
      * @return {boolean}
      */
