+++ lib/WebSocketManager.js
@@ -1,15 +1,17 @@
 const CONST = require(`../util/constants.json`);
 const EventEmitter = require(`events`);
 const WebSocketFactory = require(`./WebSocketFactory.js`);
+const randomstring = require(`randomstring`);
 
 
 /**
- *
+ * WebSocket manager. Handles the multiple WS session routine.
+ * Access and refresh tokens for each session. Authentication.
  */
 class WebSocketManager extends EventEmitter {
 
     /**
-     *
+     * Create Web Socket manager
      * @param wsURL
      */
     constructor(wsURL) {
@@ -26,7 +28,7 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Check for WS session by the key
      * @param key
      * @return {boolean}
      */
@@ -37,7 +39,7 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Set Access and Refresh token for WS session by the key
      * @param key
      * @param accessToken
      * @param refreshToken
@@ -60,7 +62,7 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Get Access and Refresh tokens of WS session by the key
      * @param key
      * @return {boolean}
      */
@@ -76,28 +78,30 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Check for authorized WS session by the key
      * @param key
      * @returns {boolean}
      */
     isAuthorized(key) {
         const me = this;
 
-        return me.authorizedKeySet.has(key);
+        return me.wsFactory.hasSocket(key) && me.authorizedKeySet.has(key);
     }
 
     /**
-     *
+     * Mark WS session by the key as authorized
      * @param key
      */
     setAuthorized(key) {
         const me = this;
 
-        me.authorizedKeySet.add(key);
+        if (me.wsFactory.hasSocket(key)) {
+            me.authorizedKeySet.add(key);
+        }
     }
 
     /**
-     *
+     * Remove authorized mark from WS session by the key
      * @param key
      */
     removeAuthorized(key) {
@@ -117,14 +121,16 @@ class WebSocketManager extends EventEmitter {
         const me = this;
 
         return me.send(key, {
-            action: CONST.WS.ACTIONS.TOKEN,
-            login: login,
-            password: password
-        }).then((tokens) => {
-            me.setTokens(key, tokens);
+                action: CONST.WS.ACTIONS.TOKEN,
+                login: login,
+                password: password
+            })
+            .then((tokens) => {
+                me.setTokens(key, tokens);
 
-            return tokens;
-        }).catch((err) => console.log(err));
+                return tokens;
+            })
+            .catch((err) => console.log(err));
     }
 
     /**
@@ -137,27 +143,59 @@ class WebSocketManager extends EventEmitter {
         const me = this;
 
         return me.send(key, {
-            action: CONST.WS.ACTIONS.AUTHENTICATE,
-            token: accessToken
-        }).then(() => {
-            me.setAuthorized(key);
-        }).catch((err) => console.log(err));
+                action: CONST.WS.ACTIONS.AUTHENTICATE,
+                token: accessToken
+            })
+            .then(() => {
+                me.setAuthorized(key);
+            })
+            .catch((err) => console.log(err));
     }
 
     /**
-     *
+     * Send message object over WS session by the key
      * @param key
-     * @param message
+     * @param messageObject
      * @return {Promise}
      */
-    send(key, message) {
+    send(key, messageObject) {
         const me = this;
 
         me._lock(key);
 
         return me.wsFactory.getSocket(key)
             .then((wSocket) => {
-                return wSocket.send(message);
+                return new Promise((resolve, reject) => {
+                    const hasRequestId = messageObject.requestId;
+
+                    if (!hasRequestId) {
+                        messageObject.requestId = randomstring.generate();
+                    }
+
+                    wSocket.sendString(JSON.stringify(messageObject));
+
+                    const listener = (event) => {
+                        const messageData = JSON.parse(event.data);
+
+                        if (messageData.action === messageObject.action) {
+                            if (messageData.requestId === messageObject.requestId) {
+                                if (!hasRequestId) {
+                                    delete messageData.requestId;
+                                }
+
+                                wSocket.removeListener(`message`, listener);
+
+                                if (messageData.status === CONST.WS.SUCCESS_STATUS) {
+                                    resolve(messageData);
+                                } else {
+                                    reject(messageData);
+                                }
+                            }
+                        }
+                    };
+
+                    wSocket.addEventListener(`message`, listener);
+                });
             })
             .then((response) => {
                 me._unlock(key);
@@ -170,7 +208,7 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Send string message (forward string message) over WS session by the key
      * @param key
      * @param stringMessage
      */
@@ -188,7 +226,7 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Close WS session by the key
      * @param key
      */
     close(key) {
@@ -204,11 +242,11 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Lock WS session resource
      * @param key
      * @private
      */
-    _lock (key) {
+    _lock(key) {
         const me = this;
 
         if (me.keyLockCounterMap.has(key)) {
@@ -225,11 +263,11 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Unlock WS session resource
      * @param key
      * @private
      */
-    _unlock (key) {
+    _unlock(key) {
         const me = this;
 
         if (me.keyLockCounterMap.has(key)) {
@@ -246,12 +284,12 @@ class WebSocketManager extends EventEmitter {
     }
 
     /**
-     *
+     * Check for locked WS session resource
      * @param key
      * @return {boolean}
      * @private
      */
-    _isLocked (key) {
+    _isLocked(key) {
         const me = this;
         let result = false;
