+++ test/integration/api/cluster.js
@@ -1,0 +1,82 @@
+const mqtt = require(`mqtt`);
+const EventEmitter = require('events');
+const randomString = require(`randomstring`);
+const sinon = require(`sinon`);
+const chai = require(`chai`);
+const expect = chai.expect;
+const assert = chai.assert;
+
+const ee = new EventEmitter();
+
+const DH_RESPONSE_TOPIC = `dh/response`;
+const DH_REQUEST_TOPIC = `dh/request`;
+const MQTT_BROKER_URL = `mqtt://localhost:1883`;
+const TEST_LOGIN = `mqtt_proxy_test_login`;
+const TEST_PASSWORD = `qwertyui`;
+const TEST_USER_ID = 14347;
+const SUCCESS_STATUS = `success`;
+const ERROR_STATUS = `error`;
+const DEVICE_ID = `VQjfBdTl0LvMVBt9RTJMOmwdqr6hWLjln1wZ`;
+const NETWORK_ID = 12276;
+const SUBJECT = `cluster`;
+const INFO_OPERATION = `info`;
+const INFO_ACTION = `${SUBJECT}/${INFO_OPERATION}`;
+const INFO_TOPIC = `${DH_RESPONSE_TOPIC}/${INFO_ACTION}`;
+let mqttClient;
+
+it(`should connect to MQTT broker`, () => {
+    return new Promise((resolve) => {
+        mqttClient = mqtt.connect(MQTT_BROKER_URL, {
+            username: TEST_LOGIN,
+            password: TEST_PASSWORD
+        });
+
+        mqttClient.on(`message`, (topic, message) => {
+            ee.emit(topic.split(`/`)[3].split(`@`)[0], JSON.parse(message.toString()))
+        });
+
+        mqttClient.on('connect', () => {
+            resolve();
+        });
+    });
+});
+
+it(`should subscribe for "${INFO_TOPIC}" topic`, () => {
+    return new Promise((resolve, reject) => {
+        mqttClient.subscribe(`${INFO_TOPIC}@${mqttClient.options.clientId}`, (err) => {
+            if (err) {
+                reject();
+            } else {
+                resolve();
+            }
+        });
+    });
+});
+
+it(`should get cluster information`, () => {
+    const requestId = randomString.generate();
+
+    return new Promise((resolve) => {
+        ee.once(INFO_OPERATION, (message) => {
+            if (message.requestId === requestId) {
+                expect(message.status).to.equal(SUCCESS_STATUS);
+                expect(message).to.include.all.keys(`clusterInfo`);
+
+                resolve();
+            }
+        });
+
+        mqttClient.publish(DH_REQUEST_TOPIC, JSON.stringify({
+            action: INFO_ACTION,
+            requestId: requestId
+        }));
+    });
+});
+
+it(`should disconnect from MQTT broker`, () => {
+    return new Promise((resolve) => {
+        mqttClient.end(() => {
+            resolve();
+        });
+    });
+});
